Update Chests & EnemyHealthbar

+++ Pixel-Throne-Rise-of-the-King/Assets/Scripts/Enemies/FlyingEnemy/FlyingEnemy.cs
@@ -35,7 +35,7 @@ public class FlyingEnemy : Enemy
         }
     }
 
-    public override void TakeDamage(int damage)
+    public override void TakeDamage(float damage)
     {
         base.TakeDamage(damage);
     }
